:sparkles: make variable regex search robust
- Improve variable regexp search by making all fields regexp rather than
  like queries
- Make all regexp matches case insensitive
- Return no results rather than an error when you search with a partial
  regexp
