+++ adminSiteServer/apiRouter.ts
@@ -20,6 +20,7 @@ import {
     getMergedGrapherConfigForVariable,
     fetchS3MetadataByPath,
     fetchS3DataValuesByPath,
+    searchVariables,
 } from "../db/model/Variable.js"
 import {
     applyPatch,
@@ -170,11 +171,11 @@ const getPostsForSlugs = async (
                     AND post_status='publish'
                     AND (
                         ${slugs
-                            .map(
-                                () =>
-                                    `post_content REGEXP CONCAT('grapher/', ?, '[^a-zA-Z_\-]')`
-                            )
-                            .join(" OR ")}
+                .map(
+                    () =>
+                        `post_content REGEXP CONCAT('grapher/', ?, '[^a-zA-Z_\-]')`
+                )
+                .join(" OR ")}
                     )
             `,
             slugs.map(lodash.escapeRegExp)
@@ -659,14 +660,14 @@ apiRouter.get(
 
         let dataset:
             | {
-                  id: number
-                  name: string
-                  version: string
-                  namespace: string
-                  isPrivate: boolean
-                  nonRedistributable: boolean
-                  variables: { id: number; name: string }[]
-              }
+                id: number
+                name: string
+                version: string
+                namespace: string
+                isPrivate: boolean
+                nonRedistributable: boolean
+                variables: { id: number; name: string }[]
+            }
             | undefined
         for (const row of rows) {
             if (!dataset || row.datasetName !== dataset.name) {
@@ -979,9 +980,9 @@ apiRouter.post(
                 ) {
                     throw new JsonError(
                         `Expected all "${obj.key}" values to be non-null and of ` +
-                            `type "${obj.expectedType}", but one or more chart ` +
-                            `configs contains a "${obj.key}" value that does ` +
-                            `not meet this criteria.`
+                        `type "${obj.expectedType}", but one or more chart ` +
+                        `configs contains a "${obj.key}" value that does ` +
+                        `not meet this criteria.`
                     )
                 }
             })
@@ -1140,9 +1141,9 @@ apiRouter.post(
                     if (!config.hasOwnProperty(k)) {
                         throw new JsonError(
                             `The "${k}" field is required, but one or more ` +
-                                `chart configs in the database does not ` +
-                                `contain it. Please report this issue to a ` +
-                                `developer.`
+                            `chart configs in the database does not ` +
+                            `contain it. Please report this issue to a ` +
+                            `developer.`
                         )
                     }
                 })
@@ -1212,9 +1213,8 @@ apiRouter.post(
             if (suggestedConfigs.length - result.affectedRows > 0) {
                 messages.push({
                     type: "warning",
-                    text: `${
-                        suggestedConfigs.length - result.affectedRows
-                    } chart revisions have not been queued for approval (e.g. because the chart revision does not contain any changes).`,
+                    text: `${suggestedConfigs.length - result.affectedRows
+                        } chart revisions have not been queued for approval (e.g. because the chart revision does not contain any changes).`,
                 })
             }
         })
@@ -1341,7 +1341,7 @@ apiRouter.post(
             if (!canUpdate) {
                 throw new JsonError(
                     `Suggest chart revision ${suggestedChartRevisionId} cannot be ` +
-                        `updated with status="${status}".`,
+                    `updated with status="${status}".`,
                     404
                 )
             }
@@ -1488,135 +1488,8 @@ apiRouter.post("/users/add", async (req: Request, res: Response) => {
 
 apiRouter.get("/variables.json", async (req) => {
     const limit = parseIntOrUndefined(req.query.limit as string) ?? 50
-    const searchStr = req.query.search as string
-
-    const whereClauses = []
-    if (searchStr) {
-        for (let part of searchStr.split(" ")) {
-            part = part.trim()
-            let not = " "
-            if (part.startsWith("-")) {
-                part = part.substring(1)
-                not = "NOT "
-            }
-            if (part.startsWith("name:")) {
-                const q = part.substring("name:".length)
-                // force a case insensitive regex search
-                q &&
-                    whereClauses.push(
-                        `${not} REGEXP_LIKE(v.name, ${escape(q)}, 'i')`
-                    )
-            } else if (part.startsWith("path:")) {
-                const q = part.substring("path:".length)
-                q &&
-                    whereClauses.push(
-                        `${not} REGEXP_LIKE(v.catalogPath, ${escape(q)}, 'i')`
-                    )
-            } else if (part.startsWith("namespace:")) {
-                const q = part.substring("namespace:".length)
-                q &&
-                    whereClauses.push(
-                        `v.catalogPath ${not} LIKE "grapher/${q}/%/%/%#%"`
-                    )
-            } else if (part.startsWith("version:")) {
-                const q = part.substring("version:".length)
-                q &&
-                    whereClauses.push(
-                        `v.catalogPath ${not} LIKE "grapher/%/${q}/%/%#%"`
-                    )
-            } else if (part.startsWith("dataset:")) {
-                const q = part.substring("dataset:".length)
-                q &&
-                    whereClauses.push(
-                        `v.catalogPath ${not} LIKE "grapher/%/%/${q}/%#%"`
-                    )
-            } else if (part.startsWith("table:")) {
-                const q = part.substring("table:".length)
-                q &&
-                    whereClauses.push(
-                        `v.catalogPath ${not} LIKE "grapher/%/%/%/${q}#%"`
-                    )
-            } else if (part.startsWith("short:")) {
-                const q = part.substring("short:".length)
-                q &&
-                    whereClauses.push(
-                        `v.catalogPath ${not} LIKE "grapher/%/%/%/%#${q}"`
-                    )
-            } else if (part.startsWith("before:")) {
-                const q = part.substring("before:".length)
-                q &&
-                    whereClauses.push(
-                        `${not} IF(d.version is not null, d.version < ${escape(
-                            q
-                        )}, cast(date(d.createdAt) as char) < ${escape(q)})`
-                    )
-            } else if (part.startsWith("after:")) {
-                const q = part.substring("after:".length)
-                q &&
-                    whereClauses.push(
-                        `${not} (IF (d.version is not null, d.version = "latest" OR d.version > ${escape(
-                            q
-                        )}, cast(date(d.createdAt) as char) > ${escape(q)}))`
-                    )
-            } else if (part === "is:published") {
-                whereClauses.push(`${not} (NOT d.isPrivate)`)
-            } else if (part === "is:private") {
-                whereClauses.push(`${not} d.isPrivate`)
-            } else {
-                part &&
-                    whereClauses.push(
-                        `${not} (REGEXP_LIKE(v.name, ${escape(
-                            part
-                        )}, 'i') OR REGEXP_LIKE(v.catalogPath, ${escape(
-                            part
-                        )}, 'i'))`
-                    )
-            }
-        }
-    }
-
-    const query = `
-        SELECT
-            v.id,
-            v.name,
-            catalogPath,
-            d.id AS datasetId,
-            d.name AS datasetName,
-            d.isPrivate AS isPrivate,
-            d.nonRedistributable AS nonRedistributable,
-            d.dataEditedAt AS uploadedAt,
-            u.fullName AS uploadedBy
-        FROM variables AS v
-        JOIN active_datasets d ON d.id=v.datasetId
-        JOIN users u ON u.id=d.dataEditedByUserId
-        ${whereClauses.length ? "WHERE " + whereClauses.join(" AND ") : ""}
-        ORDER BY d.dataEditedAt DESC
-        LIMIT ?
-    `
-    console.log(query)
-
-    const rows = await db.queryMysql(query, [limit])
-
-    const numTotalRows = (
-        await db.queryMysql(`SELECT COUNT(*) as count FROM variables`)
-    )[0].count
-
-    rows.forEach((row: any) => {
-        if (row.catalogPath) {
-            const [path, shortName] = row.catalogPath.split("#")
-            const [namespace, version, dataset, table] = path
-                .substring("grapher/".length)
-                .split("/")
-
-            row.namespace = namespace
-            row.version = version
-            row.dataset = dataset
-            row.table = table
-            row.shortName = shortName
-        }
-    })
-
-    return { variables: rows, numTotalRows: numTotalRows }
+    const query = req.query.search as string
+    return await searchVariables(query, limit)
 })
 
 apiRouter.get(
@@ -2752,8 +2625,8 @@ apiRouter.put("/gdocs/:id", async (req, res) => {
             prevGdoc.published && nextGdoc.published
                 ? "Updating"
                 : !prevGdoc.published && nextGdoc.published
-                ? "Publishing"
-                : "Unpublishing"
+                    ? "Publishing"
+                    : "Unpublishing"
         await triggerStaticBuild(res.locals.user, `${action} ${nextGdoc.slug}`)
     }
